+++ Pages/new_quiz
@@ -1,0 +1,32 @@
+from cache import CacheCat, QuizCache
+from menu import Menu_Factory
+from page import PageFactory, PageOptions, Page
+from question import Question
+from quiz import Quiz, QuizKeys
+from storage import Storage
+
+
+class NewQuiz(Page):
+
+    def display(self):
+        _name = Quiz.prompt_for_name()
+        _style = Quiz.prompt_for_style()
+        _question_file = _name + Question.extension  # location of question file determined by config file
+        _values = (_name, _style, _question_file)
+        _quiz = Quiz(_values, QuizKeys.get_keys(), True)
+        self.question_file_path = '/home/czechmate/Documents/python/programs/Quiz_Maker/data/' + _quiz.get_name() + ".qst"
+        self.store(_quiz)
+        _header = "\nWould you like to add questions?"
+        _selection = Menu_Factory.run_yes_no_menu(_header)
+        if _selection == 'no' or _selection == 'n' or _selection == 'N':
+            return PageFactory.create_page(PageOptions.home_screen)
+        else:
+            return PageFactory.create_page(PageOptions.add_questions_to_quiz, self.question_file_path)
+
+    def store(self, _quiz):
+        QuizCache.add(CacheCat.quiz, _quiz)
+        # FIXME -> Update filepath to correct path when done testing
+        _quiz_file_path = Storage.get_config_value(
+            '/home/czechmate/Documents/python/programs/Quiz_Maker/data/file_paths.txt', 'quiz_file_path')
+        Storage.append_element_to_file(_quiz_file_path, _quiz.content)
+        Storage.create_new_file(self.question_file_path)
